feat(#37): 다중 웹소켓 통합
