+++ Coconut-Backend/src/main/java/com/coconut/stock_app/dto/StockChartDTO.java
@@ -14,11 +14,14 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Builder
 public class StockChartDTO {
-    private String stockCode;
-    private String time;
-    private BigDecimal openPrice;
-    private BigDecimal highPrice;
-    private BigDecimal lowPrice;
-    private BigDecimal closePrice;
-    private Long volume;
+    private String stockCode;//0
+    private String time;//1
+    private BigDecimal currentPrice;//2
+    private BigDecimal openPrice;//7
+    private BigDecimal highPrice;//8
+    private BigDecimal lowPrice;//8
+    private String versusSign;//3
+    private BigDecimal contingentVol;//12
+    private BigDecimal accumulatedVol;//13
+    private BigDecimal accumulatedAmount; //14
 }
